feat(schema): make JsonColumnSchema library-agnostic via StandardSchema
Replace ArkType-specific Type constraint with StandardSchemaWithJSONSchema,
enabling JsonColumnSchema to work with any Standard Schema compliant library
(ArkType, Zod v4.2+, Valibot with adapter).

Changes:
- Create standard-schema.ts with StandardSchemaWithJSONSchema combined type
- Update JsonColumnSchema to use StandardSchemaWithJSONSchema constraint
- Update type inference from TSchema['infer'] to StandardSchemaV1.InferOutput
- Update columns.ts json() function signature
- Update arktype-yjs.ts and drizzle.ts converters
- Deduplicate StandardSchemaWithJSONSchema (move from actions.ts to schema)
