+++ packages/epicenter/src/core/schema/index.ts
@@ -60,6 +60,10 @@ export {
 // ============================================================================
 export { serializeCellValue } from './serialization';
 // ============================================================================
+// Standard Schema
+// ============================================================================
+export type { StandardSchemaWithJSONSchema } from './standard-schema';
+// ============================================================================
 // Types
 // ============================================================================
 export type {
